pyproject.toml
[tool.poetry]
name = "goblet-get-request"
version = "0.1.0"
description = ""
authors = ["rasnes <rasmusnes@gmail.com>"]
readme = "README.md"
packages = [{include = "goblet_get_request"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
goblet-gcp = "^0.9.2"


[tool.poetry.group.dev.dependencies]
black = {version = "^22.10.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
